requirements.txt
albumentations==1.3.0
matplotlib==3.10.6
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
peft==0.17.0
transformers==4.55.0
pycocotools==2.0.11
PyYAML==6.0.2
scikit_learn==1.7.1
timm==1.0.19
torch==2.8.0
torch_dct==0.1.6
torchvision==0.23.0
tqdm==4.67.1
